wrote artwork objects for first 2/8 artists
please check if this works. i don't understand what "portrait" means, but i interpreted it as original vs. print, where "true" is original and "false" is print. there are two objects for "please remain calm" because we have both prints and originals for that one. also, if a piece of art has multiple mediums, how is that indicated? i put a comma between but not sure if you prefer different notation. and should we include prices here as well? ty!! if this is fine i'll continue with the other pieces.